refactor: upd global styles, fix Slider logic, pull out lone components from folders
